App home improvements! (#133)
* Show an error message if the views_publish call fails

This means that if our Block Kit has a syntax error and Slack API returns en error for it, a fallback screen is shown instead of indefinite loading

* Include nav buttons in the header

* Fix header formatting error

I think this was a GH Copilot hallucination that I accepted bc it looked very similar to the OG code :(

* Show header on the loading screen

* Make clicking on a button during the loading phase work

* Delete views/home/apps.py

I don't think this view is used anywhere

* Use the new header style everywhere

* Remove buttons.py (no longer used)

* Show "dashboard" tab as selected during first load

This means that the "default" virtual view name no longer exists, but we have a DEFAULT_VIEW constant now

* Add at least some sort of return type to get_header()

* Make the helper and assigned views non-helper-friendly

* Allow non-helpers to access the app home

* Remove useless `or` statement

* Extract error_screen to a component

* Gate the tags view to only helpers

* Make the tags view non-helper-friendly

* Make the stats view non-helper-friendly

* Make the error text a bit nicer

* Fix type error

* Make the infinite loading situation less likely

This could happen if you spam-clicked one of the buttons (I tested with assigned tickets). I think the problem existed before but was made worse by the new `user_last_requested_view` logic - now it's back to normal and the issue is very hard to recreate.

* Remove "unknown user" view and transcripts

* Remove commented code block

* Remove comment the second

* Fix timezone fetching and add more logs